featt(typescript): convert test files to ts
