--- tests/__mocks__/glob.js
@@ -1,16 +1,0 @@
-'use strict'
-
-const glob = jest.genMockFromModule('glob')
-
-let mockedResults = {}
-
-module.exports = (pattern, cb) => {
-    if (pattern === '__defineMocks') {
-        mockedResults = cb
-        return
-    }
-
-    const mockedResult = mockedResults[pattern]
-    if (mockedResult === undefined) return cb(null, [])
-    cb(null, mockedResult)
-}
