+++ tests/extensions/fixtures/__mocks__/fixture.js
@@ -1,0 +1,1 @@
+module.exports = () => ({ type: 'javascript module', testing: true })
