+++ tests/extensions/fixtures/__mocks__/fixtures/js_file.js
@@ -1,0 +1,3 @@
+module.exports = () => ({
+    data: 'fixture loaded from javascript module',
+})
