+++ tests/extensions/fixtures/__mocks__/fixtures/multi.json
@@ -1,0 +1,5 @@
+{
+    "data": "fixture from JSON",
+    "testing": true,
+    "awesomeness": 100
+}
