+++ tests/extensions/fixtures/__mocks__/fixtures/multi.yaml
@@ -1,0 +1,2 @@
+type: yaml
+testing: true
