+++ tests/extensions/fixtures/__mocks__/fixtures/text_file.txt
@@ -1,0 +1,1 @@
+some text
