+++ tests/extensions/fixtures/__mocks__/fixtures/yaml_file.yaml
@@ -1,0 +1,2 @@
+type: yaml
+testing: true
