+++ tests/extensions/fixtures/__mocks__/no_default_function.js
@@ -1,0 +1,1 @@
+exports.noDefaultFunction = () => ({ type: 'javascript module', testing: true })
