+++ tests/extensions/snapshot/fixtures.js
@@ -1,4 +1,4 @@
-const dedent = require('../../../src/extensions/snapshot/dedent')
+const { dedent } = require('../../../src/extensions/snapshot/dedent')
 
 exports.featureFileContent1 = dedent`
     """
@@ -150,7 +150,13 @@ exports.diffErrorFile1WithPropertyMatchers = dedent`
     \u001b[2m  }\u001b[22m
 `
 
-exports.value1 = { key1: 'value1', key2: 'value2', key3: 'value3', key4: 'value4', key5: 'value5' }
+exports.value1 = {
+    key1: 'value1',
+    key2: 'value2',
+    key3: 'value3',
+    key4: 'value4',
+    key5: 'value5',
+}
 exports.value1WithError = {
     key1: 'value1',
     key2: 2,
@@ -158,10 +164,25 @@ exports.value1WithError = {
     key4: 'value4',
     key5: 'value5',
 }
-exports.value2 = { key1: 'value1', key2: 'value2', key3: 'value8', key4: 'value4', key5: 'value5' }
-exports.value3 = { key1: 'value1', key2: 'value2', key3: 'value9', key4: 'value4', key5: 'value5' }
+exports.value2 = {
+    key1: 'value1',
+    key2: 'value2',
+    key3: 'value8',
+    key4: 'value4',
+    key5: 'value5',
+}
+exports.value3 = {
+    key1: 'value1',
+    key2: 'value2',
+    key3: 'value9',
+    key4: 'value4',
+    key5: 'value5',
+}
 exports.multilineValue = {
-    content: { text: 'i am a text', 'long text': 'I\r\n am \r\n a\r\n long\r\n text' },
+    content: {
+        text: 'i am a text',
+        'long text': 'I\r\n am \r\n a\r\n long\r\n text',
+    },
 }
 
 exports.featureFile1 = './snapshot1.feature'
